requirements.txt
mailjet-rest
python-dotenv
streamlit
